refactor(api): move Create struct
It's been used also in Axum implementation and ActixWeb implementation
will be removed.
