+++ src/models/torrent.rs
@@ -1,7 +1,8 @@
 use serde::{Deserialize, Serialize};
 
+use super::torrent_tag::TagId;
+use crate::errors::ServiceError;
 use crate::models::torrent_file::Torrent;
-use crate::routes::torrent::Create;
 
 #[allow(clippy::module_name_repetitions)]
 pub type TorrentId = i64;
@@ -27,4 +28,27 @@ pub struct TorrentListing {
 pub struct TorrentRequest {
     pub fields: Create,
     pub torrent: Torrent,
+}
+
+#[derive(Debug, Deserialize)]
+pub struct Create {
+    pub title: String,
+    pub description: String,
+    pub category: String,
+    pub tags: Vec<TagId>,
+}
+
+impl Create {
+    /// Returns the verify of this [`Create`].
+    ///
+    /// # Errors
+    ///
+    /// This function will return an `BadRequest` error if the `title` or the `category` is empty.
+    pub fn verify(&self) -> Result<(), ServiceError> {
+        if self.title.is_empty() || self.category.is_empty() {
+            Err(ServiceError::BadRequest)
+        } else {
+            Ok(())
+        }
+    }
 }
