+++ src/web/api/v1/contexts/torrent/handlers.rs
@@ -14,9 +14,8 @@ use super::responses::{new_torrent_response, torrent_file_response};
 use crate::common::AppData;
 use crate::errors::ServiceError;
 use crate::models::info_hash::InfoHash;
-use crate::models::torrent::TorrentRequest;
+use crate::models::torrent::{Create, TorrentRequest};
 use crate::models::torrent_tag::TagId;
-use crate::routes::torrent::Create;
 use crate::services::torrent::ListingRequest;
 use crate::utils::parse_torrent;
 use crate::web::api::v1::auth::get_optional_logged_in_user;
